requirements.txt

streamlit
pandas
openai
python-docx
pdfplumber
